+++ src/main/kotlin/com/example/multidatasourceexample/service/dto/FindOrderResultDto.kt
@@ -12,7 +12,7 @@ data class FindOrderResultDto(
     val orderItems: List<FindOrderItemResultDto>,
 ) {
     companion object {
-        fun of(order: Order, isLazyLoading: Boolean = false) = FindOrderResultDto(
+        operator fun invoke(order: Order, isLazyLoading: Boolean = false) = FindOrderResultDto(
             orderId = order.id,
             memberId = order.memberId,
             statusLabel = order.status.label,
@@ -24,6 +24,24 @@ data class FindOrderResultDto(
         )
 
         private fun getFindOrderItemRessultDtos(orderItems: List<OrderItem>, isLazyLoading: Boolean) =
-            if (isLazyLoading) orderItems.map { FindOrderItemResultDto.of(orderItem = it) } else listOf()
+            if (isLazyLoading) orderItems.map { FindOrderItemResultDto(orderItem = it) } else listOf()
+    }
+}
+
+data class FindOrderItemResultDto(
+    val orderItemId: Long,
+    val categoryLabel: String,
+    val itemName: String,
+    val purchasePrice: String,
+    val createdAt: String,
+) {
+    companion object {
+        operator fun invoke(orderItem: OrderItem) = FindOrderItemResultDto(
+            orderItemId = orderItem.id,
+            categoryLabel = orderItem.category.label,
+            itemName = orderItem.itemName,
+            purchasePrice = orderItem.purchasePrice.toString(),
+            createdAt = DateTimeFormatterUtils.toStringDateTime(localDateTime = orderItem.createdAt),
+        )
     }
 }
